feat: Extract user management tools to dedicated module (Phase 6)
- Move all 5 user management tools to tools/user_management.py:
  - list_workspace_members
  - search_slack_users
  - get_user_presence
  - get_user_timezone
  - get_user_conversations

- Implement register_tools() pattern for module registration
- Update imports and registration in main server file
- Update tools/__init__.py to include user_management module
- Reduce main server file by ~395 lines (from 1236 to 841 lines)

This completes Phase 6 of the modularization refactor, maintaining
backward compatibility while improving code organization.

Fixes #22
