+++ src/slack_mcp/tools/__init__.py
@@ -4,5 +4,6 @@
 from . import core
 from . import message_management
 from . import channel_management
+from . import user_management
 
-__all__ = ["core", "message_management", "channel_management"]
+__all__ = ["core", "message_management", "channel_management", "user_management"]
